Remove code highlighting preview samples
